[Setting] 문제풀이 초기 세팅
